[skip ci] refactor sth into qn_kernel
