+++ app/src/main/java/me/singleneuron/qn_kernel/dispacher/StartActivityHook.kt
@@ -1,10 +1,11 @@
-package me.singleneuron.hook
+package me.singleneuron.qn_kernel.dispacher
 
 import android.app.Activity
 import android.content.ContextWrapper
 import android.content.Intent
 import de.robv.android.xposed.XposedBridge
 import me.singleneuron.base.adapter.BaseDelayableHookAdapter
+import me.singleneuron.hook.DebugDump
 import me.singleneuron.hook.decorator.DisableQzoneSlideCamera
 import nil.nadph.qnotified.SyncUtils
